requirements.txt
fastapi==0.123.4
uvicorn==0.38.0
httpx==0.28.1
python-dotenv==1.2.1

numpy>=2.2.6
opencv-python-headless==4.12.0.88
pillow==12.0.0

ultralytics==8.3.234
google-genai==1.53.0

python-multipart==0.0.20
pydantic==2.12.5
